@@ -4,44 +4,67 @@
   "scripts": {
     "dev": "next",
     "build": "next build",
-    "start": "next start",
-    "fix-css": "npx stylelint '**/*.scss' --fix"
+    "format": "prettier --write './**/*.{js,jsx,css,html}'",
+    "testbuild": "yarn test && yarn build",
+    "lint": "next lint",
+    "lint:js": "yarn eslint --quiet --cache --ext .js,.jsx,.ts,.tsx --format=pretty",
+    "lint:js:all": "yarn lint:js .",
+    "fix-css": "npx stylelint '**/*.css' --fix",
+    "lint:style": "stylelint '**/*.css'"
+  },
+  "lint-staged": {
+    "**/*.less": [
+      "yarn lint:style:fix"
+    ],
+    "**/*.{js,jsx,ts,tsx}": [
+      "eslint --format=pretty"
+    ],
+    "**/*.{js,jsx,tsx,ts,md,json}": [
+      "prettier --write"
+    ]
   },
   "dependencies": {
-    "@material-ui/core": "^4.10.0",
-    "@material-ui/icons": "^4.9.1",
-    "babel-plugin-import": "^1.13.0",
+    "@emotion/cache": "^11.9.3",
+    "@emotion/react": "^11.9.3",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.9.3",
+    "@mdx-js/loader": "^2.1.1",
+    "@mdx-js/react": "^2.1.1",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.8.4",
+    "@mui/styles": "^5.8.4",
+    "@next/mdx": "^12.1.6",
     "localStorage": "^1.0.4",
-    "lost": "^8.3.1",
     "next": "^12.1.6",
     "next-compose-plugins": "^2.2.0",
     "next-progressbar": "^1.0.0",
-    "now": "^18.0.0",
-    "postcss-nested": "^4.2.1",
     "react": "17.0.2",
+    "react-darkreader": "^1.5.6",
     "react-dom": "17.0.2",
-    "sass": "^1.26.8",
-    "styled-jsx-plugin-postcss": "^2.0.1"
+    "tailwindcss": "^3.1.3"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.4.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.3.0",
-    "eslint-loader": "^3.0.0",
-    "eslint-plugin-css-modules": "^2.11.0",
-    "eslint-plugin-import": "^2.18.2",
+    "@babel/core": "^7.9.0",
+    "@commitlint/cli": "^17.0.2",
+    "@commitlint/config-conventional": "^17.0.2",
+    "babel-plugin-module-resolver": "^4.0.0",
+    "babel-plugin-transform-imports": "^2.0.0",
+    "eslint": "^8.16.0",
+    "eslint-config-airbnb": "^18.1.0",
+    "eslint-config-next": "^12.1.6",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-config-react": "^1.1.7",
+    "eslint-formatter-pretty": "^4.1.0",
+    "eslint-import-resolver-babel-module": "^5.1.2",
+    "eslint-loader": "^4.0.0",
+    "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.14.3",
-    "prettier": "^1.18.2",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.19.0",
+    "husky": "^8.0.1",
+    "lint-staged": "^13.0.0",
+    "prettier": "^2.0.4",
     "stylelint": "^13.7.0",
     "stylelint-config-standard": "^20.0.0"
-  },
-  "postcss": {
-    "plugins": {
-      "lost": {},
-      "postcss-nested": {}
-    }
   }
 }
