@@ -1,63 +1,63 @@
-import { createMuiTheme } from '@material-ui/core/styles'
-import { zhCN } from '@material-ui/core/locale'
+import { createTheme } from '@mui/material/styles'
+import { zhCN } from '@mui/material/locale'
 
-export const darkTheme = createMuiTheme(
+export const darkTheme = createTheme(
   {
     palette: {
       primary: {
         main: '#FF4053',
-        contrastText: '#fff'
+        contrastText: '#fff',
       },
       secondary: {
         main: '#19857b',
-        contrastText: '#fff'
+        contrastText: '#fff',
       },
       error: {
-        main: '#556cd6'
+        main: '#556cd6',
       },
       type: 'dark',
       background: {
-        paper: '#161A2B'
+        paper: '#161A2B',
       },
       text: {
-        primary: '#fff'
+        primary: '#fff',
       },
       action: {
-        disabledBackground: '#8A8C99'
+        disabledBackground: '#8A8C99',
       },
       info: {
-        main: '#82678f'
-      }
-    }
+        main: '#82678f',
+      },
+    },
   },
   zhCN
 )
 
-export const lightTheme = createMuiTheme(
+export const lightTheme = createTheme(
   {
     palette: {
       primary: {
         main: '#FF4053',
-        contrastText: '#fff'
+        contrastText: '#fff',
       },
       secondary: {
         main: '#19857b',
-        contrastText: '#000'
+        contrastText: '#000',
       },
       type: 'light',
       text: {
-        primary: '#000'
+        primary: '#000',
       },
       background: {
-        paper: '#fff'
+        paper: '#fff',
       },
       action: {
-        disabledBackground: '#bababa'
+        disabledBackground: '#bababa',
       },
       error: {
-        main: '#556cd6'
-      }
-    }
+        main: '#556cd6',
+      },
+    },
   },
   zhCN
 )
