@@ -1,11 +1,12 @@
-import React from 'react'
+import { Children } from 'react'
 import Document, { Html, Head, Main, NextScript } from 'next/document'
-import { ServerStyleSheets } from '@material-ui/core/styles'
+import createEmotionServer from '@emotion/server/create-instance'
+import { createEmotionCache } from '../utils'
 
-export default class MyDocument extends Document {
+class CustomDocument extends Document {
   render() {
     return (
-      <Html>
+      <Html lang="en">
         <Head />
         <body>
           <Main />
@@ -16,16 +17,27 @@ export default class MyDocument extends Document {
   }
 }
 
-MyDocument.getInitialProps = async ctx => {
-  const sheets = new ServerStyleSheets()
+CustomDocument.getInitialProps = async (ctx) => {
   const originalRenderPage = ctx.renderPage
+  const cache = createEmotionCache()
+  const { extractCriticalToChunks } = createEmotionServer(cache)
+
   ctx.renderPage = () =>
     originalRenderPage({
-      enhanceApp: App => props => sheets.collect(<App {...props} />)
+      // eslint-disable-next-line
+      enhanceApp: (App) => (props) => <App emotionCache={cache} {...props} />,
     })
+
   const initialProps = await Document.getInitialProps(ctx)
+  const emotionStyles = extractCriticalToChunks(initialProps.html)
+  const emotionStyleTags = emotionStyles.styles.map((style) => (
+    <style data-emotion={`${style.key} ${style.ids.join(' ')}`} key={style.key} dangerouslySetInnerHTML={{ __html: style.css }} />
+  ))
+
   return {
     ...initialProps,
-    styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()]
+    styles: [...Children.toArray(initialProps.styles), ...emotionStyleTags],
   }
 }
+
+export default CustomDocument
